+++ src/main/java/br/com/lsukys/centraldeerros/service/ApplicationService.java
@@ -2,6 +2,6 @@ package br.com.lsukys.centraldeerros.service;
 
 import br.com.lsukys.centraldeerros.entity.Application;
 
-public interface ApplicationService extends CrudService<Application, Long> {
+public interface ApplicationService extends CrudService<Application, Integer> {
 	
 }
